pyproject.toml
[project]
name = "paste"
version = "0.1.0"
description = "A zero-friction paste service using FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "python-multipart>=0.0.17",
    "Pillow>=11.1.0",
    "markdown>=3.7",
    "pygments>=2.18.0",
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.alembic]
script_location = "alembic"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
